Task-05: move product logic to productWrapper
